Break up a word into syllables
This uses the `possible_phones` utility function to determine the
phones that compose a word, and then break those phones up into
different syllables.

The CMU dictionary marks phones with stress by ending their phone
with a number (indicating the level of stress). By building up a
phonetic spelling of the syllable, and splitting each syllable by
the stressed phone, we are able to determine roughly the number of
syllables in a word.

We return the list of syllables as the combination of the individual
phones instead of detecting where the phones lie within the words.
This will allow for consistent syllable pronunciation being returned,
which should help those relying on the syllables.