Calculate rhyming similarity between words
Basically it works against two words to determine how many phones
match in the words. It works in reverse, assuming that if two words
have the same phones in the end then they rhyme.

By default, the entire word will be compared. It will work in
reverse checking the shorter word to the longer, checking each
individual phone. The score that is returned is the number of phones
which are the same over the total number of phones in the shortest
word.

The range of phones to check can be overridden with the
`start_phone` and `end_phone` keyword arguments, which control
what sections of words are checked against each other. As Python
allows negative slices, this allows for any part of a word to be
compared, whether it is the start, middle, or end.