Started detecting rhyme schemes
The rhyme scheme is detected based on how close two lines are,
with a similarity threshold of at least 50%.

This also makes a slight change to how the similarity index between
two words is generated, so now phones with emphasis are given a
higher similarity score than phones without emphasis.  This helps
battle the previous issue of determining the similarity between
short words (like "go" and "know" or "sound").